chore: 예외처리 추가

+++ src/main/java/com/ocp/ocp_finalproject/workflow/service/WorkflowServiceImpl.java
@@ -193,6 +193,10 @@ public class WorkflowServiceImpl implements WorkflowService {
         Workflow workflow = workflowRepository.findWorkflow(userId, workflowId)
                 .orElseThrow(() -> new CustomException(WORKFLOW_NOT_FOUND));
 
+        if (workflow.getStatus() == WorkflowStatus.DELETED) {
+            throw new CustomException(ALREADY_DELETED);
+        }
+
         workflow.delete();
 
         schedulerSyncService.removeWorkflowJobs(workflowId);
